% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_ab}
\alias{calc_ab}
\title{Calculate slope and intercept}
\usage{
calc_ab(t2, t1)
}
\arguments{
\item{t2}{upper intercept}

\item{t1}{lower intercept}
}
\description{
Calculate slope and intercept
}
