% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_dens_hist}
\alias{calc_dens_hist}
\title{Calculate scaled 1d density}
\usage{
calc_dens_hist(dat, binwidth = 50, bw = 30, type = "kde",
  age_range = c(0, 4560))
}
\arguments{
\item{dat}{data.frame}

\item{binwidth}{Histogram binwidth}

\item{bw}{Density Bandwith}

\item{type}{'kde': KDE; 'pdd': detrital zircon PDD}

\item{age_range}{Age range to calculated density over}
}
\value{
Returns density
}
\description{
Calculates 1d density of age data and scales it so that it can be plotted
in the same plot of a histogram of the age data
}
