% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_dkw}
\alias{calc_dkw}
\title{Dvoretzky-Kiefer-Wolfowitz inequality}
\usage{
calc_dkw(dat, column = "age", alpha = 0.05)
}
\arguments{
\item{dat}{data.frame}

\item{column}{which column to use}

\item{alpha}{Desired alpha level}
}
\value{
data.frame with ecdf and confidence bands
}
\description{
Calculate confidence bands for ecdfs using the Dvoretzky-Kiefer-Wolfowitz
inequality.
}
\references{
Dvoretzky, A., Kiefer, J., Wolfowitz, J., 1956. Asymptotic
Minimax Character of the Sample Distribution Function and of the Classical
Multinomial Estimator. Ann. Math. Stat. 27, 642-669.
doi:10.1214/aoms/1177728174
}

