% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_o_param}
\alias{calc_o_param}
\title{Calculate 1-O}
\usage{
calc_o_param(dat1, dat2, column, alpha = 0.05, digits = 2)
}
\arguments{
\item{dat1}{data.frame}

\item{dat2}{data.frame}

\item{column}{string of name of column to use ('age' or 't_dm2')}

\item{alpha}{alpha level}

\item{digits}{number of digits}
}
\value{
1-O
}
\description{
Calculate 1-O
}
\references{
Andersen, T., Elburg, M., Cawthorn-Blazeby, A., 2015.
U-Pb and Lu-Hf zircon data in young sediments reflect sedimentary recycling
in eastern South Africa. J. Geol. Soc. London. 2006-2015.
doi:10.1144/jgs2015-006
}
