% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_p_apply}
\alias{calc_p_apply}
\title{Calculate intercepts and associated p-value}
\usage{
calc_p_apply(dat, t2, t1)
}
\arguments{
\item{dat}{data.frame}

\item{t2}{upper intercept age}

\item{t1}{lower intercept age}
}
\description{
Calculate intercepts and associated p-value
}
