% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate quantiles}
\usage{
calc_quantiles(dat, column = "t_dm2", alpha = 0.05, type = 8)
}
\arguments{
\item{dat}{data.frame}

\item{column}{which column in data.frame to use}

\item{alpha}{alpha-level (not yet used)}

\item{type}{type of quantile calculation (passed on to stats::quantile)}
}
\description{
Split up data.frame by sample-column and calculate quantiles
}
