% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{combine_matrices}
\alias{combine_matrices}
\title{Combine two square matrices}
\usage{
combine_matrices(mat1, mat2)
}
\arguments{
\item{mat1}{Matrix for upper triangle}

\item{mat2}{Matrix for lower triangle}
}
\description{
Combine two square matrices
}

