% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{dzr_mix}
\alias{dzr_mix}
\title{Calculate mixing model}
\usage{
dzr_mix(mu1, sig1, mu2, sig2)
}
\arguments{
\item{mu1}{first mean}

\item{sig1}{first standard deviation}

\item{mu2}{second mean}

\item{sig2}{second standard deviation}
}
\description{
Gaussian mixing model for detrital zircon data, using lower quantile upper
quantile plot
}
\examples{
dzr_mix(500, 50, 1000, 100)
}
