% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{find_maxima}
\alias{find_maxima}
\title{Find maxima.}
\usage{
find_maxima(dist, xmin, inc)
}
\arguments{
\item{dist}{distribution.}

\item{xmin}{minimum value of distribution.}

\item{inc}{increment.}
}
\description{
Find maxima.
}
