% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{find_plot_max}
\alias{find_plot_max}
\title{Find maximum value for plotting.}
\usage{
find_plot_max(x, accuracy = 100)
}
\arguments{
\item{x}{vector of values}

\item{accuracy}{round to nearest}
}
\description{
Find the maximum value for histogram plotting.
}
