% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{find_plot_min}
\alias{find_plot_min}
\title{Find minimum value for plotting}
\usage{
find_plot_min(x, accuracy = 100)
}
\arguments{
\item{x}{vector of values}

\item{accuracy}{round to nearest}
}
\description{
Find the minimum value for histogram plotting.
}
