% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{find_plot_min_max}
\alias{find_plot_min_max}
\title{Wrapper function for \code{find_plot_min} and \code{find_plot_max}}
\usage{
find_plot_min_max(x, accuracy = 100)
}
\arguments{
\item{x}{Age data}

\item{accuracy}{Round to nearest}
}
\value{
Returns vector of minimum and maximum plotting values
}
\description{
Find the minimum and maximum values for histogram plotting.
}
