% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{hfhf_dm}
\alias{hfhf_dm}
\title{176Lu/177Hf value of DM.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
hfhf_dm
}
\description{
176Lu/177Hf value of DM.
}
\references{
Griffin, W., Pearson, N., Belousova, E., Jackson, S., van
Achterbergh, E., O'Reilly, S. and Shee, S. 2000. The Hf isotope composition
of cratonic mantle:
LAM-MC-ICPMS analysis of zircon megacrysts in kimberlites.
Geochimica et Cosmochimica Acta 64(1), 133-147.
}
\keyword{datasets}
