% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{lambda_lu}
\alias{lambda_lu}
\title{Decay constant of 176Lu.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
lambda_lu
}
\description{
Decay constant of 176Lu.
}
\references{
Soderlund, U., Jonathan Patchett, P., Vervoort, J.D. and Isachsen
C.E. 2004. The 176Lu decay constant determined by Lu^Hf and U^Pb isotope
systematics of Precambrian mafic intrusions. Earth and Planetary Science
Letters 219, 311-324.
}
\keyword{datasets}
