% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{lambda_u235}
\alias{lambda_u235}
\title{Decay constants of 235U.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
lambda_u235
}
\description{
Decay constants of 235U.
}
\references{
Steiger, R. & Jager, E. 1977. SUBCOMMISSION ON GEOCHRONOLOGY -
CONVENTION ON USE OF DECAY CONSTANTS IN GEOCHRONOLOGY AND COSMOCHRONOLOGY.
Earth And Planetary Science Letters 36(3), 359-362.
}
\keyword{datasets}
