% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{luhf_chur}
\alias{luhf_chur}
\title{176Lu/177Hf value of CHUR.}
\format{An object of class \code{numeric} of length 1.}
\usage{
luhf_chur
}
\description{
176Lu/177Hf value of CHUR.
}
\references{
Bouvier, A., Vervoort, J.D. and Jonathan Patchett P. 2008. The
Lu-Hf and Sm-Nd isotopic composition of CHUR: Constraints from unequilibrated
chondrites and implications for the bulk composition of terrestrial planets.
Earth and Planetary Science Letters 273, 48-57.
}
\keyword{datasets}
