% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{luhf_zrc}
\alias{luhf_zrc}
\title{176Lu/177Hf value of average continental crust.}
\format{An object of class \code{numeric} of length 1.}
\usage{
luhf_zrc
}
\description{
176Lu/177Hf value of average continental crust.
}
\references{
Griffin, W., Belousova, E., Shee, S., Pearson, N. and O'Reilly,
S. 2004. Archean crustal evolution in the northern Yilgam Craton:
U-Pb and Hf-isotope evidence from detrital zircons. Precambrian Research,
231-282.
}
\keyword{datasets}
