% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{make_tiling}
\alias{make_tiling}
\title{Produce data.frame of 1-O matrix suitable for geom_tile}
\usage{
make_tiling(dat, type)
}
\arguments{
\item{dat}{data.frame}

\item{type}{What to calculate}
}
\description{
Produce data.frame of 1-O matrix suitable for geom_tile
}

