% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{Natal_group}
\alias{Natal_group}
\title{Dataset Natal group}
\description{
Detrital zircon U-Pb age and Lu-Hf data from the Natal Group, KwaZulu-Natal,
South-Africa (Kristoffersen et al. 2016).
}
\details{
Used as example of proper format of a csv-file for input in the shiny
interface.
}
\references{
Kristoffersen, M., Andersen, T., Elburg, M.A., Watkeys, M.K.,
2016. Detrital zircon in a supercontinental setting: locally derived and
far-transported components in the Ordovician Natal Group, South Africa.
J. Geol. Soc. London. 173, 203-215. doi:10.1144/jgs2015-012
}
