% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{o_param_matrix_age}
\alias{o_param_matrix_age}
\title{Populate matrix with age 1-O}
\usage{
o_param_matrix_age(dat, alpha = 0.05, digits = 3)
}
\arguments{
\item{dat}{data.frame}

\item{alpha}{alpha level}

\item{digits}{number of digits}
}
\value{
matrix of 1-O for ages
}
\description{
Populate matrix with age 1-O
}
\references{
Andersen, T., Elburg, M., Cawthorn-Blazeby, A., 2015.
U-Pb and LuHf zircon data in young sediments reflect sedimentary recycling
in eastern South Africa. J. Geol. Soc. London. 2006-2015.
doi:10.1144/jgs2015-006
}

