% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_axis_lim}
\alias{plot_axis_lim}
\title{Axes limits for ggplot2}
\usage{
plot_axis_lim(xlim = c(0, 4560), step = 200)
}
\arguments{
\item{xlim}{x-axis limit}

\item{step}{specify x-axis steps}

\item{ylim}{y-axis limit}
}
\value{
list of ggplot2::coord_cartesian object
}
\description{
Convenience function to change axes limits for ggplot2
}

