% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dens_hist}
\alias{plot_dens_hist}
\title{Plot 1d density with histogram}
\usage{
plot_dens_hist(dat, bw = 30, binwidth = 50, type = "kde",
  age_range = c(0, 4560), facet = FALSE, step = step)
}
\arguments{
\item{dat}{data.frame}

\item{bw}{density bandwidth}

\item{binwidth}{histogram binwidth}

\item{type}{'kde': traditional KDE 'pdd': detrial zircon PDD}

\item{age_range}{range over which to calculate density}

\item{facet}{logical, facet samples?}

\item{step}{specify x-axis steps}
}
\value{
ggplot2 1d density plot with histogram
}
\description{
Creates density and histogram plot of U-Pb data.
}

