% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot ecdf}
\usage{
plot_ecdf(
  dat,
  mult_ecdf = FALSE,
  column = "age",
  conf = FALSE,
  guide = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{dat}{data.frame}

\item{mult_ecdf}{logical, plot several ecdfs in same plot}

\item{column}{which column to use}

\item{conf}{logical, plot confidence bands}

\item{guide}{logical, show legend}

\item{alpha}{alpha-level used for confidence bands}
}
\description{
Plot ecdf for U-Pb age or Lu-Hf model age data.
}
