% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_labels}
\alias{plot_labels}
\title{Labels for ggplot2 plots}
\usage{
plot_labels(xlab = "Age (Ma)", ylab = "Density")
}
\arguments{
\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\value{
Returns ggplot2 labels
}
\description{
Convenience function to label ggplot2
}

