% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot quantiles}
\usage{
plot_quantiles(
  dat,
  column = "t_dm2",
  conf = FALSE,
  alpha = 0.05,
  type = 8,
  guide = TRUE,
  mix = FALSE,
  mix_data = NULL
)
}
\arguments{
\item{dat}{data.frame}

\item{column}{which column in data.frame to use}

\item{conf}{logical, plot confidence interval}

\item{alpha}{alpha-level}

\item{type}{type of quantile calculation (passed on to stats::quantile)}

\item{guide}{logical, show legend?}

\item{mix}{logical, add mixing model}

\item{mix_data}{mixing model data}
}
\description{
Plot quantiles
}
