% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_reimink}
\alias{plot_reimink}
\title{Plot likelihood of intercept ages}
\usage{
plot_reimink(dat, step)
}
\arguments{
\item{dat}{data.frame}

\item{step}{Chord step to use in intercept calculation}
}
\description{
Plot likelihood of intercept ages
}
