% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_text_options}
\alias{plot_text_options}
\title{Modify text options of plots}
\usage{
plot_text_options(font_name = "Helvetica", title_size = 10,
  label_size = 7, legend_size = 10, strip_text_y_size = 8)
}
\arguments{
\item{font_name}{Name of font to use}

\item{title_size}{Font size of x- and y-axis titles}

\item{label_size}{Font size of x- and y-axis tick labels}

\item{legend_size}{Font size of legend}

\item{strip_text_y_size}{Font size of vertical panel text}
}
\description{
Modify text options of plots
}

