% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tile}
\alias{plot_tile}
\title{Tile plot of 1-O matrix}
\usage{
plot_tile(dat, column)
}
\arguments{
\item{dat}{data.frame}

\item{column}{Column to use; 'age' or 't_dm2}
}
\description{
Tile plot of 1-O matrix
}

