% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{populate_matrix}
\alias{populate_matrix}
\title{Populate matrix}
\usage{
populate_matrix(dat, FUN, ...)
}
\arguments{
\item{dat}{data.frame}

\item{FUN}{Function used to populate matrix}

\item{...}{Additional parameters passed to function}
}
\value{
Populated matrix
}
\description{
Populate matrix
}
