% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{quant_bounds}
\alias{quant_bounds}
\title{Calculate confidence bands for lower and upper quartile}
\usage{
quant_bounds(dat, column = "t_dm2", alpha = 0.05)
}
\arguments{
\item{dat}{data.frame}

\item{column}{column to use for calculations}

\item{alpha}{alpha-level}
}
\description{
Calculate confidence bands for lower and upper quartile
}

