% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{satkoski_1d}
\alias{satkoski_1d}
\title{Calculate 1d likeness of detrital zircon populations}
\usage{
satkoski_1d(x, y, bw = 30, digits = 3)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{bw}{bandwidth}

\item{digits}{number, round result to significant digits}
}
\description{
Calculates the likeness of detrital zircon populations in 1 dimension after
Satkoski et al. (2013).
}
\references{
Satkoski, A.M., Wilkinson, B.H., Hietpas, J., Samson, S.D., 2013.
Likeness among detrital zircon populations - An approach to the comparison of
age frequency data in time and space. GSA Bulletin 125, 1783-1799.
}
