% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{satkoski_1d_matrix}
\alias{satkoski_1d_matrix}
\title{Pairwise Satkoski likeness}
\usage{
satkoski_1d_matrix(dat, bw = 30, digits = 3)
}
\arguments{
\item{dat}{data.frame}

\item{bw}{density bandwidth}

\item{digits}{number, round result to significant digits}
}
\description{
Populate a matrix with pairwise Satkoski 1d likeness.
}
\references{
Satkoski, A.M., Wilkinson, B.H., Hietpas, J., Samson, S.D., 2013.
Likeness among detrital zircon populations - An approach to the comparison of
age frequency data in time and space. GSA Bulletin 125, 1783-1799.
}

