% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{satkoski_2d}
\alias{satkoski_2d}
\title{Calculate 2d (age and Lu-Hf) likeness of detrital zircon populations}
\usage{
satkoski_2d(x, y, bw = c(30, 2.5), digits = 3)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{bw}{vector of density bandwidths}

\item{digits}{number, round result to significant digits}
}
\description{
Calculates the likeness of detrital zircon populationsin 2 dimensions after
Satoski et al. (2013).
}
\references{
Satkoski, A.M., Wilkinson, B.H., Hietpas, J., Samson, S.D., 2013.
Likeness among detrital zircon populations - An approach to the comparison of
age frequency data in time and space. GSA Bulletin 125, 1783-1799.
}

