% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{u238_u235_ratio}
\alias{u238_u235_ratio}
\title{Atomic ratio of 238U and 235U.}
\format{An object of class \code{numeric} of length 1.}
\usage{
u238_u235_ratio
}
\description{
Atomic ratio of 238U and 235U.
}
\references{
Steiger, R. & Jager, E. 1977. SUBCOMMISSION ON GEOCHRONOLOGY -
CONVENTION ON USE OF DECAY CONSTANTS IN GEOCHRONOLOGY AND COSMOCHRONOLOGY.
Earth And Planetary Science Letters 36(3), 359-362.
}
\keyword{datasets}

