% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{check_conc}
\alias{check_conc}
\title{Check concordancy of input ages}
\usage{
check_conc(dat, disc_lim = 10)
}
\arguments{
\item{dat}{data.frame containing at least ages and percentage of discordancy}

\item{disc_lim}{Discordancy limit}
}
\value{
Concordant data
}
\description{
Check the concordancy of the U-Pb data and return the data within the desired
discordancy limit.
}
