% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{make_tiling}
\alias{make_tiling}
\title{Produce data.frame of 1-O matrix suitable for geom_tile}
\usage{
make_tiling(dat, column)
}
\arguments{
\item{dat}{data.frame}

\item{column}{Column to use; 'age' or 't_dm2'}
}
\description{
Produce data.frame of 1-O matrix suitable for geom_tile
}

