% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dens}
\alias{plot_dens}
\title{Plot 1d density}
\usage{
plot_dens(
  dat,
  bw = 30,
  type = "kde",
  age_range = c(0, 4560),
  facet = FALSE,
  fixed_y = FALSE,
  step = 200
)
}
\arguments{
\item{dat}{data.frame}

\item{bw}{density bandwidth}

\item{type}{'kde': traditional KDE 'pdd': detrital zircon PDD}

\item{age_range}{range over which to calculate density}

\item{facet}{logical, facet samples?}

\item{fixed_y}{logical, fixed y-axis?}

\item{step}{specify x-axis steps}
}
\value{
ggplot2 1d density plot with histogram
}
\description{
Creates density for U-Pb data.
}
