% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hf}
\alias{plot_hf}
\title{Plot Lu-Hf data}
\usage{
plot_hf(dat, range = c(0, 4560), plot_type = "ehf", guide = TRUE,
  contours = FALSE, contour_data = NULL, combine_contours = FALSE,
  constants)
}
\arguments{
\item{dat}{data.frame}

\item{range}{range to display}

\item{plot_type}{'ehf'=epsilon-Hf; 'hfhf'=176/177Hf}

\item{guide}{logical, show legend?}

\item{contours}{logical, plot contours?}

\item{contour_data}{data.frame containing data to contour}

\item{combine_contours}{logical combine contouring data}

\item{constants}{vector of constants which must be in the order
decay constant 176Lu, 176/177Hf CHUR, 176Lu/177Hf CHUR, 176/177Hf DM and
176Lu/177Hf DM}
}
\description{
Plot Lu-Hf data as both epsilon-Hf vs. age and 176/177Hf vs. age.
}
